Refactor subject service and controller and resolved adding course

+++ src/main/java/com/equipeor/isepu/repository/ProfessorRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Collection;
 import java.util.Optional;
 
-public interface ProfessorRepository extends JpaRepository<Professor,Integer> {
-     Optional<Professor> findById(long id);
+public interface ProfessorRepository extends JpaRepository<Professor,Long> {
+     Optional<Professor> findById(Long id);
      Collection<Professor> findByCourses_Subject_Name(String name);
      boolean existsByEmail(String email);
      Optional<Professor> findByEmail(String email);
